// apps/api/src/common/enums/video.enum.ts
export enum VideoSiteEnum {
  YOUTUBE = 'YouTube',
  VIMEO = 'Vimeo',
}

export enum VideoTypeEnum {
  TRAILER = 'Trailer',
  TEASER = 'Teaser',
  CLIP = 'Clip',
  FEATURETTE = 'Featurette',
  BEHIND_THE_SCENES = 'Behind the Scenes',
  BLOOPERS = 'Bloopers',
}

export enum VideoLanguageEnum {
  EN = 'en',
  UK = 'uk',
}